Refactor repository layout
