+++ pkg/plugins/gifs/gifs.go
